Fix container view position when used with animation and from a storyboard.
When created using storyboard, the property tabBarFrame is CGSizeZero. It was also wrongly repositioning viewContainer frame, not considering tabor size correctly. The error was caught when tab bar is above container view, and the size (width and/or height) of the tab bar plus container view is less then the size of the view controller that contains both.